requirements.txt
djangorestframework
psycopg2-binary
python-dotenv
stripe